Debuged NDEigensystem
NDEigensystem_1D_4 is the last stable notebook.

The names of the control parameters of the manipulates were chenaged. Now the manipulates are independent and they don't conflict.

multi_segment_1D_1: still at the experimental level, it need more work about the boundary matching and its relations with tenergy and full normalization of the wave function.